Simple ET keyboard V1.0
Working, very simple (single line) eye-tracking based speller.
The files include GazeStream, which must be launched before the application itself in order to make it work. The code is written in Processing 3, which is required to run, at the moment, but an executable version can be easily created within Processing (I just could not upload it here because of the number of files involved).